requirements.txt
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
Markdown==2.6.8
pyFirmata==1.0.3
pyserial==3.3
pytz==2017.2
